add User model, routes, controller, and tests for creating a user, password hashing and validation

+++ src/database/pre-seed.ts
@@ -1,10 +1,19 @@
 import Todo from '../models/Todo.model';
+import User from '../models/User.model';
 
-// Used to remove all todos before testing begins
-const preSeedFunction = () => {
-	return Todo.destroy({
-		where: {},
-	});
+// Used to remove all todos and users before testing begins
+const preSeedFunction = (data: string) => {
+	if (data === 'todo') {
+		return Todo.destroy({
+			where: {},
+		});
+	}
+
+	if (data === 'user') {
+		return User.destroy({
+			where: {},
+		});
+	}
 };
 
 export default preSeedFunction;
